chore: rename tree utility files
- jsonTreeUtils.js → treeUtils.js
- jsonInteractiveControls.js → interactiveControls.js
- Update import paths accordingly

Related to camunda/camunda-modeler#5953
