+++ lib/components/CodeMirrorEditor/interactiveControls.js
