+++ lib/components/CodeMirrorEditor/treeUtils.js
