+++ lib/hooks/useContextMenuBehavior.js
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useCallback } from 'react';
-import { buildPathFromSyntaxNode, extractValue, getLineContext } from '../components/CodeMirrorEditor/jsonTreeUtils';
+import { buildPathFromSyntaxNode, extractValue, getLineContext } from '../components/CodeMirrorEditor/treeUtils';
 import useTracking from './useTracking';
 
 export default function useContextMenuBehavior({ menuState, view, rootVariableName, onClose }) {
